Let IMAP authentication immediately fail when no password was given
Some clients (old Outlook) try IMAP authentication on first connection
without a password, when it was not saved in the client beforehand.
So to support scenarios where the client tries to login with IMAP on
first connection to email server and OAuth 2.0 authorisation didn't
happen yet, the authorisation will be delayed until the correct
password is entered in the client. As this password will be used for
token encryption/decryption in config file, other protocols, like SMTP,
which also needs a password to work, can now use the same password to
decrypt the authorisation token.